Do not Fatal when we get context deadline for BtrDB
Currently due to some slower drives on the archiver server we are using,
we sometimes get this context deadline. The solution is NOT to increase
the deadline because it is already at 30 seconds and that is plenty
reasonable. For now lets just trust that Btrdb will do its best to input
the record, so we don't handle the deadline error. Retrying immediately
is probably not the right thing to do because we are more likely to get
this error during periods of high load, so we don't want to place more
load on the database when that happens. Likely that we will revisit this
design decision in the future