Prefiltering on points. Switching over to int64 timestamps
There's a lot in this commit! Dotmaster has had its time range methods
expanded with Intersections for time ranges and disjoint ranges. This is
used to prefilter a data query so that instead of querying for all the
data a user has asked for and then iterating through those points to
filter out those from times that are inaccessible by that user, we
simply rewrite the user's query so its in the bounds of what they're
allowed to see.

The danger of doing this is that we lose out an some natural parallelism
if we dispatch a bunch of UUIDs to BtrDB and ask for the same range for
all of them; this is potentially a bit faster (and we could introduce a
special case for that), but this is already netting us a 30-50% increase
in query speed, which is great